refactor: 사다리 결과와 관련된 도메인을 별도의 패키지로 분리한다
